+++ tests/testthat/test-get_loglikelihood.R
@@ -36,11 +36,16 @@ test_that("get_loglikelihood - lm", {
   expect_equal(as.numeric(get_loglikelihood(x)), 18.4205, tolerance = 1e-3)
   expect_equal(as.numeric(get_loglikelihood(x, check_response = TRUE)), -75.58669, tolerance = 1e-3)
 
-
   # sqrt-response
   x <- lm(sqrt(mpg) ~ wt, data = mtcars)
   expect_equal(as.numeric(get_loglikelihood(x)), -7.395031, tolerance = 1e-3)
   expect_equal(as.numeric(get_loglikelihood(x, check_response = TRUE)), -76.89597, tolerance = 1e-3)
+
+  # power to x
+  x <- lm(mpg^3.5 ~ wt, weights = wg, data = mtcars)
+  expect_equal(as.numeric(get_loglikelihood(x)), -385.6256, tolerance = 1e-3)
+  expect_equal(as.numeric(get_loglikelihood(x, check_response = TRUE)), -110.5192, tolerance = 1e-3)
+
 })
 
 test_that("get_loglikelihood - glm", {